+++ src/main/java/domain/CardDeck.java
@@ -1,37 +1,21 @@
 package domain;
 
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 
 public class CardDeck {
-    private static final List<TrumpCard> CARD_DECK = new ArrayList<>();
 
-    static {
-        for (CardShape cardShape : CardShape.values()) {
-            for (CardNumber cardNumber : CardNumber.values()) {
-                CARD_DECK.add(new TrumpCard(cardShape, cardNumber));
-            }
-        }
-        Collections.shuffle(CARD_DECK);
+    private final List<TrumpCard> cards;
+
+    public CardDeck() {
+        this.cards = new ArrayList<>();
     }
 
-    public static TrumpCard drawCard() {
-        if (CARD_DECK.isEmpty()) {
-            throw new IllegalArgumentException("카드가 다 떨어졌습니다");
-        }
-        return CARD_DECK.removeFirst();
+    public void addTrumpCard(TrumpCard card) {
+        cards.add(card);
     }
 
-    public static TrumpCard getCard(int index) {
-        return CARD_DECK.get(index);
-    }
-
-    public static List<TrumpCard> getCardDeck() {
-        return CARD_DECK;
-    }
-
-    public int size() {
-        return 0;
+    public int cardsSize() {
+        return cards.size();
     }
 }
